# nodes/arm/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The arm API for ArmOS to interact with a variety of different robotic arms, starting with the AR3. It uses the basic interface of `x,y,z,a,b,c` for control.",
        "title": "ArmOS arm API",
        "contact": {},
        "version": "0.1"
    },
    "basePath": "/api/",
    "paths": {
        "/calibrate": {
            "post": {
                "description": "Calibrates the robot. Should be done occasionally to affirm the robot is where we think it should be.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "low_level"
                ],
                "summary": "Calibrate the arm",
                "parameters": [
                    {
                        "description": "joints to calibrate",
                        "name": "joints",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CalibrateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/directions": {
            "get": {
                "description": "Returns current direction of arm's motors.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setup"
                ],
                "summary": "Returns direction of arm joints",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.JointDirections"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/move": {
            "post": {
                "description": "Moves the robot's stepper motors.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "low_level"
                ],
                "summary": "Move the arm's stepper motors",
                "parameters": [
                    {
                        "description": "steppers coordinates",
                        "name": "move",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.MoveStepperInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "dev"
                ],
                "summary": "A pingable endpoint",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/set_directions": {
            "post": {
                "description": "Sets up the robots joints. This only has to be done once during the setup of the robot.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "setup"
                ],
                "summary": "Sets direction of arm joints",
                "parameters": [
                    {
                        "description": "direction of joints",
                        "name": "directions",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.JointDirections"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.CalibrateInput": {
            "type": "object",
            "properties": {
                "j1": {
                    "type": "boolean"
                },
                "j2": {
                    "type": "boolean"
                },
                "j3": {
                    "type": "boolean"
                },
                "j4": {
                    "type": "boolean"
                },
                "j5": {
                    "type": "boolean"
                },
                "j6": {
                    "type": "boolean"
                },
                "speed": {
                    "type": "integer"
                },
                "tr": {
                    "type": "boolean"
                }
            }
        },
        "main.JointDirections": {
            "type": "object",
            "properties": {
                "j1": {
                    "type": "boolean"
                },
                "j2": {
                    "type": "boolean"
                },
                "j3": {
                    "type": "boolean"
                },
                "j4": {
                    "type": "boolean"
                },
                "j5": {
                    "type": "boolean"
                },
                "j6": {
                    "type": "boolean"
                },
                "tr": {
                    "type": "boolean"
                }
            }
        },
        "main.MoveStepperInput": {
            "type": "object",
            "properties": {
                "accdur": {
                    "type": "integer"
                },
                "accspd": {
                    "type": "integer"
                },
                "dccdur": {
                    "type": "integer"
                },
                "dccspd": {
                    "type": "integer"
                },
                "j1": {
                    "type": "integer"
                },
                "j2": {
                    "type": "integer"
                },
                "j3": {
                    "type": "integer"
                },
                "j4": {
                    "type": "integer"
                },
                "j5": {
                    "type": "integer"
                },
                "j6": {
                    "type": "integer"
                },
                "speed": {
                    "type": "integer"
                },
                "tr": {
                    "type": "integer"
                }
            }
        }
    }
}